Desktop/Scorpius1.2-main/backend/bridge/requirements.txt
# Scorpius Bridge Network Dependencies

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0  # PostgreSQL async driver
sqlmodel>=0.0.14

# Redis and caching
redis==4.5.4
hiredis>=2.2.0

# Blockchain libraries
web3>=6.11.0
eth-account>=0.9.0
solana>=0.30.0
py-solc-x>=1.12.0
websockets==11.0.3

# Cryptography and security
cryptography>=41.0.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6

# Monitoring and observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
structlog>=23.2.0
python-json-logger>=2.0.0

# Message queues and events
kafka-python>=2.0.0
celery>=5.3.0
kombu>=5.3.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
httpx==0.23.3
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Deployment
docker>=6.1.0
gunicorn>=21.2.0

# Configuration and environment
python-dotenv>=1.0.0
toml>=0.10.0
pyyaml>=6.0.0

# Utilities
click>=8.1.0
rich>=13.7.0
typer>=0.9.0
aiofiles>=23.2.0
